+++ src/server/game/Maps/PhaseMgr.h
@@ -48,7 +48,7 @@ struct PhaseDefinition
 {
     uint32 zoneId;
     uint32 entry;
-    uint32 phasemask;
+    uint64 phasemask;
     uint32 phaseId;
     uint32 terrainswapmap;
     uint8 flags;
